+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [v1.1.12] - 2025-10-15
+
+### Added
+
+- Adds SMAP string variables as excluded variables list in the
+`earthdata-varinfo` configuration file. An exception is thrown when an excluded
+variable is explicitly requested.
+
 ## [v1.1.11] - 2025-10-15
 
 ### Changed
@@ -154,6 +162,7 @@ Repository structure changes include:
 For more information on internal releases prior to NASA open-source approval,
 see legacy-CHANGELOG.md.
 
+[v1.1.12]: https://github.com/nasa/harmony-opendap-subsetter/releases/tag/1.1.12
 [v1.1.11]: https://github.com/nasa/harmony-opendap-subsetter/releases/tag/1.1.11
 [v1.1.10]: https://github.com/nasa/harmony-opendap-subsetter/releases/tag/1.1.10
 [v1.1.9]: https://github.com/nasa/harmony-opendap-subsetter/releases/tag/1.1.9